+++ hosted/aks/p0/p0_importing_test.go
@@ -4,42 +4,28 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters/aks"
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
 	"github.com/rancher/rancher/tests/framework/pkg/config"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/aks/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var _ = Describe("P0Importing", func() {
-	var (
-		ctx         helpers.Context
-		clusterName string
-		location    = "eastus"
-		k8sVersion  = "1.26.6"
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("akshostcluster")
-		ctx = helpers.CommonBeforeSuite("aks")
-	})
+
 	When("a cluster is imported", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
 			err = helper.CreateAKSClusterOnAzure(location, clusterName, k8sVersion, "1")
 			Expect(err).To(BeNil())
 			aksConfig := new(helper.ImportClusterConfig)
 			config.LoadAndUpdateConfig(aks.AKSClusterConfigConfigurationFileKey, aksConfig, func() {
 				aksConfig.ResourceGroup = clusterName
-				aksConfig.ResourceLocation = location
 			})
+
 			cluster, err = helper.ImportAKSHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 			Expect(err).To(BeNil())
 			cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -48,7 +34,7 @@ var _ = Describe("P0Importing", func() {
 			cluster.AKSConfig = cluster.AKSStatus.UpstreamSpec
 		})
 		AfterEach(func() {
-			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherClient)
+			err = helper.DeleteAKSHostCluster(cluster, ctx.RancherClient)
 			Expect(err).To(BeNil())
 			err = helper.DeleteAKSClusteronAzure(clusterName)
 			Expect(err).To(BeNil())
@@ -79,7 +65,6 @@ var _ = Describe("P0Importing", func() {
 			initialNodeCount := *cluster.AKSConfig.NodePools[0].Count
 
 			By("scaling up the nodepool", func() {
-				var err error
 				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
 				Expect(err).To(BeNil())
 				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -90,7 +75,6 @@ var _ = Describe("P0Importing", func() {
 			})
 
 			By("scaling down the nodepool", func() {
-				var err error
 				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
 				Expect(err).To(BeNil())
 				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -101,7 +85,6 @@ var _ = Describe("P0Importing", func() {
 			})
 
 			By("adding a nodepool/s", func() {
-				var err error
 				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
 				Expect(err).To(BeNil())
 				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -109,7 +92,6 @@ var _ = Describe("P0Importing", func() {
 				Expect(len(cluster.AKSConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
 			})
 			By("deleting the nodepool", func() {
-				var err error
 				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
 				Expect(err).To(BeNil())
 				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -130,7 +112,6 @@ var _ = Describe("P0Importing", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the ControlPlane", func() {
-					var err error
 					cluster, err = helper.UpgradeClusterKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -142,7 +123,6 @@ var _ = Describe("P0Importing", func() {
 				})
 
 				By("upgrading the NodePools", func() {
-					var err error
 					cluster, err = helper.UpgradeNodeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
