+++ hosted/aks/p0/p0_provisioning_test.go
@@ -6,33 +6,20 @@ import (
 
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
+	"github.com/rancher/rancher/tests/framework/pkg/config"
 
-	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters/aks"
-	"github.com/rancher/rancher/tests/framework/pkg/config"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/aks/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var _ = Describe("P0Provisioning", func() {
-	var (
-		clusterName string
-		ctx         helpers.Context
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("akshostcluster")
-		ctx = helpers.CommonBeforeSuite("aks")
 
-	})
 	When("a cluster is created", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
 			aksConfig := new(aks.ClusterConfig)
 			config.LoadAndUpdateConfig(aks.AKSClusterConfigConfigurationFileKey, aksConfig, func() {
 				aksConfig.ResourceGroup = clusterName
@@ -50,7 +37,7 @@ var _ = Describe("P0Provisioning", func() {
 			err = helper.DeleteAKSClusteronAzure(clusterName)
 			Expect(err).To(BeNil())
 		})
-		It("should successfully provision the cluster", func() {
+		It("should successfully provision the cluster & add, delete, scale nodepool", func() {
 
 			By("checking cluster name is same", func() {
 				Expect(cluster.Name).To(BeEquivalentTo(clusterName))
@@ -72,6 +59,43 @@ var _ = Describe("P0Provisioning", func() {
 				Expect(podErrors).To(BeEmpty())
 			})
 
+			currentNodePoolNumber := len(cluster.AKSConfig.NodePools)
+			initialNodeCount := *cluster.AKSConfig.NodePools[0].Count
+
+			By("adding a nodepool", func() {
+				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.AKSConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
+			})
+			By("deleting the nodepool", func() {
+				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.AKSConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
+			})
+
+			By("scaling up the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.AKSConfig.NodePools {
+					Expect(*cluster.AKSConfig.NodePools[i].Count).To(BeNumerically("==", initialNodeCount+1))
+				}
+			})
+
+			By("scaling down the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.AKSConfig.NodePools {
+					Expect(*cluster.AKSConfig.NodePools[i].Count).To(BeNumerically("==", initialNodeCount))
+				}
+			})
 		})
 
 		Context("Upgrading K8s version", func() {
@@ -86,7 +110,6 @@ var _ = Describe("P0Provisioning", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the ControlPlane", func() {
-					var err error
 					cluster, err = helper.UpgradeClusterKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -98,7 +121,6 @@ var _ = Describe("P0Provisioning", func() {
 				})
 
 				By("upgrading the NodePools", func() {
-					var err error
 					cluster, err = helper.UpgradeNodeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -108,55 +130,6 @@ var _ = Describe("P0Provisioning", func() {
 						Expect(np.OrchestratorVersion).To(BeEquivalentTo(upgradeToVersion))
 					}
 				})
-			})
-		})
-
-		It("should be possible to add or delete the nodepools", func() {
-			currentNodePoolNumber := len(cluster.AKSConfig.NodePools)
-
-			By("adding a nodepool", func() {
-				var err error
-				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.AKSConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
-			})
-			By("deleting the nodepool", func() {
-				var err error
-				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.AKSConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
-
-			})
-
-		})
-
-		It("should be possible to scale up/down the nodepool", func() {
-			initialNodeCount := *cluster.AKSConfig.NodePools[0].Count
-
-			By("scaling up the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.AKSConfig.NodePools {
-					Expect(*cluster.AKSConfig.NodePools[i].Count).To(BeNumerically("==", initialNodeCount+1))
-				}
-			})
-
-			By("scaling down the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.AKSConfig.NodePools {
-					Expect(*cluster.AKSConfig.NodePools[i].Count).To(BeNumerically("==", initialNodeCount))
-				}
 			})
 		})
 	})
