+++ hosted/eks/p0/p0_importing_test.go
@@ -1,39 +1,22 @@
 package p0_test
 
 import (
-	"os"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters"
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/eks/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var _ = Describe("P0Importing", func() {
-	var (
-		clusterName string
-		ctx         helpers.Context
-		region      = os.Getenv("EKS_REGION")
-		k8sVersion  = "1.26"
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("ekshostcluster")
-		ctx = helpers.CommonBeforeSuite("eks")
-	})
 
 	When("a cluster is imported", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
 			err = helper.CreateEKSClusterOnAWS(region, clusterName, k8sVersion, "1")
 			Expect(err).To(BeNil())
 			cluster, err = helper.ImportEKSHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
@@ -51,7 +34,7 @@ var _ = Describe("P0Importing", func() {
 			// TODO: Force delete EKS cluster
 		})
 
-		It("should successfully import the cluster", func() {
+		It("should successfully import the cluster & add, delete, scale nodepool", func() {
 
 			By("checking cluster name is same", func() {
 				Expect(cluster.Name).To(BeEquivalentTo(clusterName))
@@ -73,6 +56,45 @@ var _ = Describe("P0Importing", func() {
 				Expect(podErrors).To(BeEmpty())
 			})
 
+			currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
+			initialNodeCount := *cluster.EKSConfig.NodeGroups[0].DesiredSize
+
+			By("scaling up the NodeGroup", func() {
+				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount+1)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.EKSConfig.NodeGroups {
+					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount+1))
+				}
+			})
+
+			By("scaling down the NodeGroup", func() {
+				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.EKSConfig.NodeGroups {
+					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount))
+				}
+			})
+
+			By("adding a NodeGroup/s", func() {
+				cluster, err = helper.AddNodeGroup(cluster, increaseBy, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber+1))
+			})
+			By("deleting the NodeGroup", func() {
+				cluster, err = helper.DeleteNodeGroup(cluster, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber))
+
+			})
+
 		})
 
 		Context("Upgrading K8s version", func() {
@@ -86,7 +108,6 @@ var _ = Describe("P0Importing", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the ControlPlane", func() {
-					var err error
 					cluster, err = helper.UpgradeClusterKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -95,7 +116,6 @@ var _ = Describe("P0Importing", func() {
 				})
 
 				By("upgrading the NodeGroups", func() {
-					var err error
 					cluster, err = helper.UpgradeNodeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					// TODO Does not upgrades version, since using custom LT
@@ -105,54 +125,6 @@ var _ = Describe("P0Importing", func() {
 					// 	Expect(ng.Version).To(BeEquivalentTo(upgradeToVersion))
 					// }
 				})
-			})
-		})
-
-		It("should be possible to add or delete the NodeGroups", func() {
-			currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
-
-			By("adding a NodeGroup/s", func() {
-				var err error
-				cluster, err = helper.AddNodeGroup(cluster, increaseBy, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber+1))
-			})
-			By("deleting the NodeGroup", func() {
-				var err error
-				cluster, err = helper.DeleteNodeGroup(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber))
-
-			})
-
-		})
-		It("should be possible to scale up/down the NodeGroup", func() {
-			initialNodeCount := *cluster.EKSConfig.NodeGroups[0].DesiredSize
-
-			By("scaling up the NodeGroup", func() {
-				var err error
-				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount+1)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.EKSConfig.NodeGroups {
-					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount+1))
-				}
-			})
-
-			By("scaling down the NodeGroup", func() {
-				var err error
-				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.EKSConfig.NodeGroups {
-					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount))
-				}
 			})
 		})
