+++ hosted/eks/p0/p0_provisioning_test.go
@@ -4,33 +4,20 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters/eks"
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/eks/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var _ = Describe("P0Provisioning", func() {
-	var (
-		clusterName string
-		ctx         helpers.Context
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("ekshostcluster")
-		ctx = helpers.CommonBeforeSuite("eks")
-	})
 
 	When("a cluster is created", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
 			cluster, err = eks.CreateEKSHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 			Expect(err).To(BeNil())
 			cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -42,7 +29,7 @@ var _ = Describe("P0Provisioning", func() {
 			Expect(err).To(BeNil())
 		})
 
-		It("should successfully provision the cluster", func() {
+		It("should successfully provision the cluster & add, delete, scale nodepool", func() {
 
 			By("checking cluster name is same", func() {
 				Expect(cluster.Name).To(BeEquivalentTo(clusterName))
@@ -64,6 +51,45 @@ var _ = Describe("P0Provisioning", func() {
 				Expect(podErrors).To(BeEmpty())
 			})
 
+			currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
+			initialNodeCount := *cluster.EKSConfig.NodeGroups[0].DesiredSize
+
+			By("scaling up the NodeGroup", func() {
+				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount+1)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.EKSConfig.NodeGroups {
+					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount+1))
+				}
+			})
+
+			By("scaling down the NodeGroup", func() {
+				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.EKSConfig.NodeGroups {
+					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount))
+				}
+			})
+
+			By("adding a NodeGroup", func() {
+				cluster, err = helper.AddNodeGroup(cluster, increaseBy, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber+1))
+			})
+			By("deleting the NodeGroup", func() {
+				cluster, err = helper.DeleteNodeGroup(cluster, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber))
+
+			})
+
 		})
 
 		Context("Upgrading K8s version", func() {
@@ -77,7 +103,6 @@ var _ = Describe("P0Provisioning", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the ControlPlane", func() {
-					var err error
 					cluster, err = helper.UpgradeClusterKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -86,7 +111,6 @@ var _ = Describe("P0Provisioning", func() {
 				})
 
 				By("upgrading the NodeGroups", func() {
-					var err error
 					cluster, err = helper.UpgradeNodeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -95,54 +119,6 @@ var _ = Describe("P0Provisioning", func() {
 						Expect(ng.Version).To(BeEquivalentTo(upgradeToVersion))
 					}
 				})
-			})
-		})
-
-		It("should be possible to add or delete the NodeGroups", func() {
-			currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
-
-			By("adding a NodeGroup", func() {
-				var err error
-				cluster, err = helper.AddNodeGroup(cluster, increaseBy, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber+1))
-			})
-			By("deleting the NodeGroup", func() {
-				var err error
-				cluster, err = helper.DeleteNodeGroup(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber))
-
-			})
-
-		})
-		It("should be possible to scale up/down the NodeGroup", func() {
-			initialNodeCount := *cluster.EKSConfig.NodeGroups[0].DesiredSize
-
-			By("scaling up the NodeGroup", func() {
-				var err error
-				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount+1)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.EKSConfig.NodeGroups {
-					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount+1))
-				}
-			})
-
-			By("scaling down the NodeGroup", func() {
-				var err error
-				cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.EKSConfig.NodeGroups {
-					Expect(*cluster.EKSConfig.NodeGroups[i].DesiredSize).To(BeNumerically("==", initialNodeCount))
-				}
 			})
 		})
