+++ hosted/gke/p0/p0_importing_test.go
@@ -1,47 +1,30 @@
 package p0_test
 
 import (
-	"os"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/gke/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
-	management "github.com/rancher/rancher/tests/framework/clients/rancher/generated/management/v3"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters"
 	"github.com/rancher/rancher/tests/framework/extensions/clusters/gke"
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
 	"github.com/rancher/rancher/tests/framework/pkg/config"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 )
 
 var _ = Describe("P0Importing", func() {
-	var (
-		clusterName string
-		ctx         helpers.Context
-		zone        = "us-central1-c"
-		project     = os.Getenv("GKE_PROJECT_ID")
-		k8sVersion  = "1.27.4-gke.900"
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("gkehostcluster")
-		ctx = helpers.CommonBeforeSuite("gke")
-	})
 
 	When("a cluster is created", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
+			err = helper.CreateGKEClusterOnGCloud(zone, clusterName, project, k8sVersion)
+			Expect(err).To(BeNil())
+
 			gkeConfig := new(helper.ImportClusterConfig)
 			config.LoadAndUpdateConfig(gke.GKEClusterConfigConfigurationFileKey, gkeConfig, func() {
 				gkeConfig.ProjectID = project
 			})
-			err = helper.CreateGKEClusterOnGCloud(zone, clusterName, project, k8sVersion)
-			Expect(err).To(BeNil())
 			cluster, err = helper.ImportGKEHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 			Expect(err).To(BeNil())
 			cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -56,7 +39,7 @@ var _ = Describe("P0Importing", func() {
 			Expect(err).To(BeNil())
 		})
 
-		It("should successfully import the cluster", func() {
+		It("should successfully import the cluster & add, delete, scale nodepool", func() {
 
 			By("checking cluster name is same", func() {
 				Expect(cluster.Name).To(BeEquivalentTo(clusterName))
@@ -78,6 +61,45 @@ var _ = Describe("P0Importing", func() {
 				Expect(podErrors).To(BeEmpty())
 			})
 
+			currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
+			initialNodeCount := *cluster.GKEConfig.NodePools[0].InitialNodeCount
+
+			By("scaling up the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.GKEConfig.NodePools {
+					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount+1))
+				}
+			})
+
+			By("scaling down the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.GKEConfig.NodePools {
+					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount))
+				}
+			})
+
+			By("adding a nodepool", func() {
+				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
+			})
+			By("deleting the nodepool", func() {
+				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
+
+			})
+
 		})
 		Context("Upgrading K8s version", func() {
 			var upgradeToVersion *string
@@ -90,7 +112,6 @@ var _ = Describe("P0Importing", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the Controlplane & NodePools", func() {
-					var err error
 					cluster, err = helper.UpgradeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient, true)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -104,53 +125,5 @@ var _ = Describe("P0Importing", func() {
 			})
 		})
 
-		It("should be possible to add or delete the nodepools", func() {
-			currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
-
-			By("adding a nodepool", func() {
-				var err error
-				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
-			})
-			By("deleting the nodepool", func() {
-				var err error
-				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
-
-			})
-
-		})
-
-		It("should be possible to scale up/down the nodepool", func() {
-			initialNodeCount := *cluster.GKEConfig.NodePools[0].InitialNodeCount
-
-			By("scaling up the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.GKEConfig.NodePools {
-					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount+1))
-				}
-			})
-
-			By("scaling down the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.GKEConfig.NodePools {
-					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount))
-				}
-			})
-		})
 	})
 })
