+++ hosted/gke/p0/p0_provisioning_test.go
@@ -1,8 +1,6 @@
 package p0_test
 
 import (
-	"os"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
@@ -12,33 +10,21 @@ import (
 	nodestat "github.com/rancher/rancher/tests/framework/extensions/nodes"
 	"github.com/rancher/rancher/tests/framework/extensions/workloads/pods"
 	"github.com/rancher/rancher/tests/framework/pkg/config"
-	namegen "github.com/rancher/rancher/tests/framework/pkg/namegenerator"
 
 	"github.com/rancher/hosted-providers-e2e/hosted/gke/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var _ = Describe("P0Provisioning", func() {
-	var (
-		clusterName string
-		project     = os.Getenv("GKE_PROJECT_ID")
-		ctx         helpers.Context
-		increaseBy  = 1
-	)
-	var _ = BeforeEach(func() {
-		clusterName = namegen.AppendRandomString("gkehostcluster")
-		ctx = helpers.CommonBeforeSuite("gke")
-	})
 
 	When("a cluster is created", func() {
-		var cluster *management.Cluster
 
 		BeforeEach(func() {
-			var err error
 			gkeConfig := new(management.GKEClusterConfigSpec)
 			config.LoadAndUpdateConfig(gke.GKEClusterConfigConfigurationFileKey, gkeConfig, func() {
 				gkeConfig.ProjectID = project
 			})
+
 			cluster, err = gke.CreateGKEHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 			Expect(err).To(BeNil())
 			cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
@@ -48,7 +34,7 @@ var _ = Describe("P0Provisioning", func() {
 			err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
 			Expect(err).To(BeNil())
 		})
-		It("should successfully provision the cluster", func() {
+		It("should successfully provision the cluster & add, delete, scale nodepool", func() {
 
 			By("checking cluster name is same", func() {
 				Expect(cluster.Name).To(BeEquivalentTo(clusterName))
@@ -70,6 +56,45 @@ var _ = Describe("P0Provisioning", func() {
 				Expect(podErrors).To(BeEmpty())
 			})
 
+			currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
+			initialNodeCount := *cluster.GKEConfig.NodePools[0].InitialNodeCount
+
+			By("scaling up the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.GKEConfig.NodePools {
+					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount+1))
+				}
+			})
+
+			By("scaling down the nodepool", func() {
+				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				for i := range cluster.GKEConfig.NodePools {
+					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount))
+				}
+			})
+
+			By("adding a nodepool", func() {
+				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
+			})
+			By("deleting the nodepool", func() {
+				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
+				Expect(err).To(BeNil())
+				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+				Expect(err).To(BeNil())
+				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
+
+			})
+
 		})
 		Context("Upgrading K8s version", func() {
 			var upgradeToVersion, currentVersion *string
@@ -83,7 +108,6 @@ var _ = Describe("P0Provisioning", func() {
 
 			It("should be able to upgrade k8s version of the cluster", func() {
 				By("upgrading the ControlPlane", func() {
-					var err error
 					cluster, err = helper.UpgradeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient, false)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -95,7 +119,6 @@ var _ = Describe("P0Provisioning", func() {
 				})
 
 				By("upgrading the NodePools", func() {
-					var err error
 					cluster, err = helper.UpgradeKubernetesVersion(cluster, upgradeToVersion, ctx.RancherClient, true)
 					Expect(err).To(BeNil())
 					err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
@@ -109,53 +132,5 @@ var _ = Describe("P0Provisioning", func() {
 			})
 		})
 
-		It("should be possible to add or delete the nodepools", func() {
-			currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
-
-			By("adding a nodepool", func() {
-				var err error
-				cluster, err = helper.AddNodePool(cluster, increaseBy, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
-			})
-			By("deleting the nodepool", func() {
-				var err error
-				cluster, err = helper.DeleteNodePool(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber))
-
-			})
-
-		})
-
-		It("should be possible to scale up/down the nodepool", func() {
-			initialNodeCount := *cluster.GKEConfig.NodePools[0].InitialNodeCount
-
-			By("scaling up the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.GKEConfig.NodePools {
-					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount+1))
-				}
-			})
-
-			By("scaling down the nodepool", func() {
-				var err error
-				cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount)
-				Expect(err).To(BeNil())
-				err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-				Expect(err).To(BeNil())
-				for i := range cluster.GKEConfig.NodePools {
-					Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically("==", initialNodeCount))
-				}
-			})
-		})
 	})
 })
